Properly validate PUT requests
As it stands, the request body of PUT requests is simply ignored.

This stems from the reliance on the Symfony ServerRequestInterface/getParsedBody() function, which, in case of a PUT request, will simply ignore the body. Of course, that's the intended behaviour of that function, but that is not the behaviour expected in the contest of this validator, and in fact, can create a very serious security risk due to being relatively easy to miss.

As such, I offer a hotfix, in the form of the added function parse_raw_http_request() (taken from stackoverflow, all credit to the author, thread is linked), which parses body parameters in case of a PUT request, and using it in that case.

This is just a hotfix, as the author of that function warned it might not cover some edge cases, but in my opinion it's better than having no validation of PUT body parameters at all.