Fix string-to-boolean casting
It is very common that boolean value casted to "0" or "1" value. And it is the industry-standard in PHP because it is how http_build_query() cast boolean. 
I faced with many such issues and have to disable validation or use artificial string in the specification when dealing with boolean flags in the GET parameters. 
So I create this small PR to fix this issue.