Handle path parameters for arrays correctly
The parameters for serialized arrays was
using the same deserialization check.
There are different formats for header,
query and path. Split up the validation
of arrays into two different sets to
allow different validation for path
and header/query parameters.

The changes will validate path parameters
for arrays but not objects which is still
not working.